+++ .env.template
@@ -1,0 +1,16 @@
+###############
+# OPEN AI API #
+###############
+OPENAI_API_KEY=
+
+#########
+# MONGO #
+#########
+MONGO_INITDB_ROOT_USERNAME=
+MONGO_INITDB_ROOT_PASSWORD=
+MONGO_INITDB_DATABASE=
+
+#########
+# REDIS #
+#########
+REDIS_BROKER_URL=
