+++ TODO.md
@@ -1,0 +1,16 @@
+- dockerize
+- mongodb connection
+- model of flashcard
+- parsing document
+- logging
+- settings
+- ci/cd
+
+
+[x] <- docker compose: app + mongo + celery worker & redis
+[x] <- mongodb: connect
+[ ] <- tests
+[ ] <- celery: create task
+[x] <- pyproject: adapt
+[ ] <- endpoint for web's healthcheck
+[ ] <- configure dockerignore
