+++ docker-compose.yml
@@ -26,6 +26,19 @@ services:
       start_period: 20s
       timeout: 10s
 
+  celery:
+    build:
+      context: .
+      dockerfile: ./ci/dockerfiles/Dockerfile
+    entrypoint: ./ci/entrypoint.sh run-celery
+    env_file:
+      - .env
+    restart: always
+    networks:
+      - app-network
+    depends_on:
+      - redis
+
   mongodb:
     image: mongo:7.0
     env_file:
@@ -42,6 +55,8 @@ services:
 
   redis:
     image: redis:7-alpine
+    networks:
+      - app-network
 
 networks:
   app-network:
